Read message contents "correctly"
As weird as it is (maybe there's something I'm missing) adding an extra space
after the message number makes the message readable. So either this is just
something really weird, or there's a meaning to this.

Also, note that the 0 before the message contents means that the message is not
private.

Example of incoming global message from user "Bob" with contents "This is a message!":

	inMessage="ping<EOM>201 Bob 0 This is a message! <EOM>"
	inMessage="ping"