Android Studio project initialized and java src organizing
This commit adds a basic Android Studio project under the android directory.
Here is where all of the Android resources and project configurations will go.

All java source code will be under the java directory. This will make accessing
code easier because it will be so close to the root. Note that the java package
names have been modified to be the following:

	edu.wright.crowningkingsandroid for android related files
	edu.wright.crowningkingsbase for base related files

To run the client as it did for Project1, go into the java directory and use the
following command (for linux):

	javac edu/wright/crowningkingsbase/main.java
	java edu.wright.crowningkingsbase.main

Also, If you plan on using Android Studio for your IDE, make sure to add a link
to the java directory for the source at android/app/src/main/java. I have
included a script that will add this link. Just make sure to run the script from
the root of the git repo.

As of now, Project1 still exists. Whenever everyone is ready we can remove it to
clean up the reop.