More message handling, ServerMessage execution prep, basic improvements
This adds more functionality for the ServerMessageHandler, ServerMessage, and
ServerConnection classes. Also, a new UserInterface.java file has been added for
future use.

For running on desktop, the MainTest now calls the run() method of the
BaseClient class. This will help us in the future when we want to get things
working for Android.

An script "run.sh" has also been added. It just makes running the client from
the terminal a little easier.

Comments at the beginning of each file have some overall ideas of what the
class does and will do in the future. Please read these to get an overall vision
of what the class is responsible for.