Public and Private messages now send correctly
This commit fixes some issues with sending public/private messages. The issue
was with the formatting of the string in AbstractServerMessage which added a
space before the string. The server saw this space before the message code and
would reject the message. This commit fixes this bug, and also fixes the
parameters in MessageAll and MessageClient to include 0 (for public) and 1
(for private) in the string that gets sent to the server.