Add more functionality to client base
Some more general functionality has been added to crowningkings base. Some of
the new functionality includes the following:
	Joining a table, leaving a table, making a table, and quitting.
The issue that was fixed in commit b09a8aaf9d5dc1b6e26adaaf1e1687e510febb6f has
now been added to the other Server messages. This issue causes messages being
sent to the server to be improperly read if the message code is not preceded
with TWO spaces. For all current server messages, the message code
(parameter[0]) has an extra space following. This seems to fix the issues with
sending messages to the server.