Adds TableList and NewTable messages
When the client gets a TableList or NewTable message from the server, it will
get the respected message and call it's run method. The run method on these
messages will call methods in the BaseClient, which will call the ui to update
it's views.

In AbstractUserInterface, two methods were added. So make sure to add those
two methods in the desktop UI.

I also removed AndroidUI.java because I'm currently using MainActivity.java in
place of AndroidUI. I will push the updates I've made to MainActivity shortly.