Cleaned up the BaseClient class
I organized the methods in BaseClient so that they are grouped base on if they
are used to send messages to the server, or to update the ui.

Please, try to keep future created methods in the correct group. It will make
reading and understanding the code much easier in the future :)

Removed two duplicate methods that I added in the previous commit.