Android UI update
This adds multiple things to the android ui including the nav drawer for sending
private/public messages. As of now, it keeps track of tables, and when new
clients join the game. but NOT the clients already joined when the android
joined. That will be fixed shortly.

As of now, when it receives the Send Username, it always replies with AndroidBob