Clicking on a user/public message starts new message screen activity
This adds a bunch of classes from PieMessage for visualization and such.
As of now, no messages are being sent or received correctly. And the username
does not show in the message activity. This will be the next step in getting
messages to be displayed.

PieMessage can be found at https://github.com/bboyairwreck/PieMessage