BaseClient method renaming, ServerMessageHandler fixes
In this commit I changed some of the methods in the BaseClient class to more
accurately reflect the naming of the server messages. For example, addTable is
now newTable. I also reordered the methods so that they are in numerical order
based off of the message code of the server message that will call that method.

The ServerMessageHandler interpreter cases have been fixed to return the proper
ServerMessage rather than the TestMessage. But all of the messages are NOT yet
handled. There are still a few that need to be handled in the future.