Added TableActivity to be started when user joins/observes a table
When a user clicks on a table from the list, that table's status will be
displayed in a dialog box, from there the user can observe or join. doing so
sends the command to the server. When the server replies with the message that
tells the client that they are at a table, the TableActivity gets started.
The TableActivity is where the table will be shown to the user.