New AndroidUIService, Private and Public messaging now working
Since not all of the UI messages from the server have to do with the Lobby, or
any other particular activity by itself, it would be best for an underlying
service that would decide which activity would need information about which UI
changes. Because of this a service that implements the AbstractUserInterface has
been added. This service gets all of the messages and broadcasts the changes via
intents (Android's message passing system through the OS). This makes having to
have instances of each activity no longer an issue. Each activity will receive
Only the intents that matter to that activity.

Sending and receiving private and public messages through the MessageActivity is
now functional. When you see a username in the nav drawer, you can click on
their name to start the messaging activity. Then you can type out a message and
send it. Also, if you're in that activity, and the other client sends you a
message, it will show up in the activity. So you can have conversations publicly
or privately. BUT as of now, when you leave that activity all of the messages
are lost. This is something that I might work on in the future.