Android Table Activity beginnings, and basic code clean up
This just adds some basic functionality to the TableActivity. Now, when the
current table's status is changed (from another player joining) a dialog box
opens asking if the user is ready to play. This isn't fully working, but it is
conceptually there.

Other basic code cleaning was done to make all code consistent and well
formatted.