Start/Ready button on TableActivity
A Start button has been added to the TableActivity class. When a user
first joins a table this button is visible to press. When pressed, a
dialog box asks for confirmation to make sure that the user is ready.
If the user presses "YES" the start button dissappears so they can't
send another READY command. When the Android client gets the GAME_START
command from the server, it makes the board visible.