Android almost complete Added piece moving, login screen, error consts
Pieces are able to be moved when it is the client's turn. When opponent
makes move, screen does not refresh until after user clicks something.
When starting CrowningKings, you will be placed in a login activity
where the user must enter a valid username.

Other various changes have been implemented.