Bug fix for the index out of bounds exception
When something goes wrong connection with the server, the client would
continue trying to get server messages. This would spam the client with
indexOutOfBounds exceptions because the server would be sending a -1 for
the size of the message. This commit adds checking for the index, and if
it is not within the bounds it will wait a little while before getting
more messages (to stop the spamming).

Also, some small delays have been added to state change commands (such
as when client is now at lobby, table, etc.). This helps keep the state
of the ui accurate, and keeps from having messages being "handled"
during ui activity/state changes.